Fix coverage drops
Add a test to cover part of cli.py that was not being tested.

Have existing tests checking the returncode

Make the tox.ini work locally

Remove broken download badge

Detect if the interpreter given is a virtualenv and use the parent interpreter to generate the new virtualenv.

Add missing docstring